+++ scripts/WalledTemplate.js
@@ -90,17 +90,19 @@ export class WalledTemplate {
       lightWall: this.options.lastReflectedEdge // Only used for cones
     };
 
-    // Add in elevation for Wall Height to use
-    const origin = this.origin;
-    origin.b = this.origin.z;
-    origin.t = this.origin.z;
-    origin.object = {};
-
     let sweepClass = this.sweepClass;
     if ( sweepClass === LightWallSweep && !this.options.lastReflectedEdge) sweepClass = ClockwiseSweepShape;
 
     const sweep = new sweepClass();
-    sweep.initialize(origin, cfg);
+    sweep.initialize(this.origin, cfg);
+
+    // Add in elevation for Wall Height to use
+    // Default to treating template as infinite in vertical directions
+    // Do this after initialization b/c something is flipping them around. Likely Wall Height.
+    cfg.source.object ??= {};
+    cfg.source.object.b ??= Number.POSITIVE_INFINITY;
+    cfg.source.object.t ??= Number.NEGATIVE_INFINITY;
+
     sweep.compute();
     return sweep;
   }
@@ -301,7 +303,7 @@ export class WalledTemplateCircle extends WalledTemplate {
    * Get the original version of this shape.
    * @returns {PIXI.Circle}
    */
-  getOriginalShape() { return MeasuredTemplate.getCircleShape(this.distance); }
+  getOriginalShape() { return CONFIG.MeasuredTemplate.objectClass.getCircleShape(this.distance); }
 
   /**
    * Get boundary shape for this sized circle set to the origin.
@@ -310,7 +312,7 @@ export class WalledTemplateCircle extends WalledTemplate {
   getBoundaryShape() {
     // Pad the circle by one pixel so it better covers expected grid spaces.
     // (Rounding tends to drop spaces on the edge.)
-    return MeasuredTemplate.getCircleShape(this.distance + 1);
+    return CONFIG.MeasuredTemplate.objectClass.getCircleShape(this.distance + 1);
   }
 
   /**
@@ -390,7 +392,7 @@ export class WalledTemplateRectangle extends WalledTemplateCircle {
    * Get the original version of this rectangle shape.
    * @returns {PIXI.Circle}
    */
-  getOriginalShape() { return MeasuredTemplate.getRectShape(this.direction, this.distance); }
+  getOriginalShape() { return CONFIG.MeasuredTemplate.objectClass.getRectShape(this.direction, this.distance); }
 
   /**
    * Get boundary shape for this rectangle set to the origin.
@@ -504,7 +506,9 @@ export class WalledTemplateRay extends WalledTemplate {
    * Get the original version of this shape.
    * @returns {PIXI.Polygon}
    */
-  getOriginalShape() { return MeasuredTemplate.getRayShape(this.direction, this.distance, this.width); }
+  getOriginalShape() {
+    return CONFIG.MeasuredTemplate.objectClass.getRayShape(this.direction, this.distance, this.width);
+  }
 
   /**
    * Generate a new WalledTemplateRay based on reflecting off the first wall encountered
@@ -601,7 +605,9 @@ export class WalledTemplateCone extends WalledTemplateRay {
    * Get the original version of this shape.
    * @returns {PIXI.Polygon}
    */
-  getOriginalShape() { return MeasuredTemplate.getConeShape(this.direction, this.width, this.distance); }
+  getOriginalShape() {
+    return CONFIG.MeasuredTemplate.objectClass.getConeShape(this.direction, this.width, this.distance);
+  }
 
   /**
    * Locate wall segments that the cone hits.
@@ -749,11 +755,6 @@ export class WalledTemplateCone extends WalledTemplateRay {
 
       const shadowConeV = Rr.normalize().multiplyScalar(reflectionDist);
       const shadowConeOrigin = reflectionPoint.subtract(shadowConeV);
-
-//       if ( Number.isNaN(shadowConeOrigin.x) || Number.isNaN(shadowConeOrigin.x) ) {
-//         console.error("shadowConeOrigin fail!", this);
-//       }
-
 
       // Shallow copy the options for the new template.
       const opts = { ...this.options };